requirements.txt
gradio==4.44.1
insightface==0.7.3
opencv-python-headless==4.9.0.80
onnxruntime==1.18.0
onnx==1.16.0
protobuf==4.25.3
numpy==1.26.4
Pillow==10.3.0
tqdm==4.66.4
